we_are_social_pre_for/requirements/base.txt
appdirs==1.4.3
arrow==0.10.0
beautifulsoup4==4.6.0
disqus==0.0.4
Django==1.11.4
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-paypal==0.3.6
django-registration==2.2
django-tinymce==2.6.0
djangorestframework==3.6.3
funcsigs==1.0.2
gunicorn==19.7.1
mock==2.0.0
pbr==3.1.1
Pillow==4.1.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
six==1.10.0
stripe==1.55.2
git+https://github.com/hcurtis1/reusable_blog